Fixed the silly type conversion bug where convert("number",13,10,16) = 00000000000000000000000000000000000000000000000000000000000000000000000D. due to floating point errors becasuse the quotient was a double but now it's a long

+++ QuickMaffs/Convert/NumberBaseConversions.cs
@@ -73,7 +73,7 @@ namespace QuickMaffs.Convert
                 return Roman.ToRoman(number);
 
             string returnValue = "";
-            double Quotient = number;
+            long Quotient = number;
 
             while (true)
             {
